requirements.txt
openai>=2.30.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.80.0
beautifulsoup4>=4.11.0
gspread>=6.0.0
